Add files via upload
Team Name -HackOps

Title -Digital complaint box for college students.

The Digital Complaint Box is an online platform for college students to safely and easily submit complaints or feedback. It allows for anonymous or identified submissions and ensures concerns are addressed promptly by the appropriate authorities, promoting transparency and a better campus environment.